chore: split tsconfig type package regressions
