+++ crates/vize/tests/check_reference_types_cli.rs
@@ -193,107 +193,6 @@ if (import.meta.vitest) {
 }
 
 #[test]
-fn check_loads_compiler_options_types_from_tsconfig_ambient_declarations() {
-    let Some(corsa_path) = resolve_test_corsa_path() else {
-        return;
-    };
-    let project_root = unique_case_dir("compiler-options-types");
-    let _ = std::fs::remove_dir_all(&project_root);
-    std::fs::create_dir_all(project_root.join("src")).unwrap();
-    link_workspace_vue(&project_root).unwrap();
-    write(
-        &project_root,
-        "node_modules/@nuxtjs/composition-api/package.json",
-        r#"{ "types": "index.d.ts" }"#,
-    );
-    write(
-        &project_root,
-        "node_modules/@nuxtjs/composition-api/index.d.ts",
-        r#"export interface UseContextReturn {}
-export function useContext(): UseContextReturn;
-"#,
-    );
-    write(
-        &project_root,
-        "node_modules/nuxt-i18n/package.json",
-        r#"{ "types": "index.d.ts" }"#,
-    );
-    write(
-        &project_root,
-        "node_modules/nuxt-i18n/index.d.ts",
-        r#"import "@nuxtjs/composition-api";
-
-declare module "@nuxtjs/composition-api" {
-  interface UseContextReturn {
-    app: {
-      i18n: {
-        t(key: string): string;
-      };
-    };
-  }
-}
-"#,
-    );
-    write(
-        &project_root,
-        "tsconfig.json",
-        r#"{
-  "compilerOptions": {
-    "strict": true,
-    "target": "ES2022",
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "noEmit": true,
-    "types": ["nuxt-i18n"]
-  },
-  "include": ["src/**/*"]
-}"#,
-    );
-    write(
-        &project_root,
-        "src/App.vue",
-        r#"<script setup lang="ts">
-import { useContext } from "@nuxtjs/composition-api";
-
-const label: string = useContext().app.i18n.t("ready");
-void label;
-</script>
-
-<template><div /></template>
-"#,
-    );
-
-    let output = Command::new(env!("CARGO_BIN_EXE_vize"))
-        .current_dir(&project_root)
-        .env("CORSA_PATH", &corsa_path)
-        .args([
-            "check",
-            "--tsconfig",
-            "tsconfig.json",
-            "src/App.vue",
-            "--format",
-            "json",
-        ])
-        .output()
-        .unwrap();
-
-    let stdout = std::string::String::from_utf8(output.stdout).unwrap();
-    let stderr = std::string::String::from_utf8(output.stderr).unwrap();
-    assert!(
-        output.status.success(),
-        "check failed:\nstdout:\n{stdout}\nstderr:\n{stderr}"
-    );
-    let json: serde_json::Value = serde_json::from_str(&stdout).unwrap();
-    assert_eq!(json["errorCount"], serde_json::json!(0), "{stdout}");
-    assert!(
-        !stdout.contains("TS2339"),
-        "compilerOptions.types should provide module augmentations:\n{stdout}"
-    );
-
-    let _ = std::fs::remove_dir_all(&project_root);
-}
-
-#[test]
 fn check_provides_define_art_for_standalone_musea_art_files() {
     let Some(corsa_path) = resolve_test_corsa_path() else {
         return;
